requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
asgiref==3.9.1
attrs==23.2.0
Automat==22.10.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.12.3
build==1.3.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
constantly==23.10.4
cryptography==42.0.7
cssselect==1.2.0
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
dnspython==2.6.1
durationpy==0.10
fastapi==0.116.1
filelock==3.14.0
flatbuffers==25.2.10
frozenlist==1.4.1
fsspec==2025.7.0
google-api-core==2.19.0
google-api-python-client==2.179.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-secret-manager==2.20.0
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
herokuify-scrapyd==1.0
hf-xet==1.1.8
html2text==2025.4.15
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.0
httpx-sse==0.4.1
huggingface-hub==0.34.4
humanfriendly==10.0
hyperlink==21.0.0
idna==3.7
importlib_metadata==8.7.0
importlib_resources==6.5.2
incremental==22.10.0
itemadapter==0.9.0
itemloaders==1.3.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
json5==0.9.25
jsonpatch==1.33
jsonpointer==2.4
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.30
langchain-text-splitters==0.3.9
langsmith==0.4.15
lxml==5.2.2
markdown-it-py==4.0.0
markdownify==1.2.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.0.5
mypy_extensions==1.1.0
networkx==3.5
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
onnxruntime==1.22.1
openai==1.100.2
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-asgi==0.57b0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-util-http==0.57b0
orjson==3.10.3
overrides==7.7.0
packaging==23.2
pandas==2.3.2
parsel==1.9.1
pillow==11.3.0
posthog==6.6.1
Protego==0.3.1
proto-plus==1.23.0
protobuf==4.25.8
pulsar-client==3.8.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyDispatcher==2.0.7
Pygments==2.19.2
pyOpenSSL==24.1.0
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
queuelib==1.7.0
regex==2024.5.15
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rsa==4.9
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
scrapeops_scrapy==0.5.4
Scrapy==2.11.2
scrapyd==1.4.3
scrapyd-client==1.2.3
sentence-transformers==5.1.0
service-identity==24.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
starlette==0.47.2
sympy==1.14.0
tenacity==8.3.0
threadpoolctl==3.6.0
tiktoken==0.7.0
tld==0.13
tldextract==5.1.2
tokenizers==0.21.4
torch==2.8.0
tqdm==4.66.4
transformers==4.55.3
triton==3.4.0
Twisted==24.3.0
typer==0.16.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.1
tzlocal==5.2
uberegg==0.1.1
uritemplate==4.2.0
urllib3==2.2.1
uvicorn==0.35.0
uvloop==0.21.0
w3lib==2.1.2
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.3
yarl==1.9.4
youtube-transcript-api==1.2.2
zipp==3.23.0
zope.interface==6.4.post2
zstandard==0.24.0
